requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
fastapi==0.110.1
langchain==0.1.16
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
langsmith==0.1.51
numpy==1.26.4
openai==1.23.6
pydantic==1.10.15
pymongo==4.6.3
python-dotenv==1.0.1
pytz==2024.1
tiktoken==0.6.0
tqdm==4.66.2
uvicorn==0.29.0